refactor app.vue to add router-view and move main layout to home.vue. Fixed router
