+++ web-client/src/main.ts
@@ -13,10 +13,12 @@ import { faHeart } from "@fortawesome/free-regular-svg-icons";
 library.add(faHeart, faHeartS);
 
 const app = createApp(App);
+
+app.use(router);
 const pinia = createPinia();
 
 app.use(pinia);
-app.use(router);
+
 //app.provide("authStore", useAuthStore());
 app.component("font-awesome-icon", FontAwesomeIcon);
 app.mount("#app");
