+++ web-client/src/router.ts
@@ -1,14 +1,12 @@
-import { createRouter } from "vue-router";
-import { createWebHistory } from "vue-router";
-import AppVue from "./App.vue";
+import { createRouter, createWebHistory } from "vue-router";
 import CallbackVue from "./components/common/Callback.vue";
-import EditUserProfileFormVue from "./components/common/EditUserProfileForm.vue";
+import HomeVue from "./components/layout/Home.vue";
 
 const router = createRouter({
   history: createWebHistory(),
   routes: [
-    { path: "/", component: AppVue },
-    { path: "/home", component: CallbackVue, meta: { requiresAuth: true } },
+    { path: "/", component: HomeVue },
+    { path: "/callback", component: CallbackVue },
   ],
 });
